pyproject.toml
[tool.poetry]
name = "hackernews"
version = "0.1.0"
description = ""
authors = ["Roy Shemtov <royshemtov13@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
sqlmodel = "^0.0.16"
pydantic-settings = "^2.2.1"
psycopg2-binary = "^2.9.9"
email-validator = "^2.1.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"
pytest = "^8.0.2"
alembic = "^1.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
